Add command to erase a table recreating it
This feature is going to erase a table by deleting it and recreating it.

From the DynamoDB Guidelines for Working with Tables documentation:
> Deleting an entire table is significantly more efficient than removing
items one-by-one, which essentially doubles the write throughput as you
do as many delete operations as put operations.

When the table is in a stack it is going to deploy the stack without
the table and then redeploy with the table.