Refactor erase table
It was added two new classes that represent a simple table
(that was created alone probably through the console) and
a table inside a stack (a table that was created through CloudFormation).

With these two tables, it was possible to reduce code repeated
in some places. Also, if there is a bug when deleting a simple
table or a table in CloudFormation now it is easier to debug.